Added links analysis to motif_roles()
All these clearly need tests. On one side, I should compare how my code performs relative to Alyssa's code, because it is quite different (it was a bit of a mess...). On the other side, we need to do proper tests like the ones for the nodes analysis because Alyssa don't have tests for hers